fix(gates): split the barrier census rows out for the file cap
