+++ crates/perry-runtime/src/gc/barrier/mod.rs
@@ -2001,30 +2001,4 @@ pub(super) use leaf::{decode_raw_pointer_candidate, inline_slot_store_on_cached_
 pub(super) use super::barrier_store::{barrier_child_prologue, barrier_remembering_active};
 pub use maintenance::*;
 
-/// `PERRY_GC_CENSUS`: remembered-set / dirty-page tables.
-pub(crate) fn barrier_tables_census() -> Vec<crate::gc::census::SideTableRow> {
-    use crate::gc::census::{map_bytes, set_bytes, vec_bytes};
-    let mut rows = Vec::new();
-    DIRTY_OLD_PAGES.with(|s| {
-        let s = s.borrow();
-        rows.push(("gc.dirty_old_pages", s.len(), set_bytes(&s)));
-    });
-    EXTERNAL_DIRTY_SLOT_PAGES.with(|m| {
-        let m = m.borrow();
-        let inner: usize = m.values().map(vec_bytes).sum();
-        rows.push((
-            "gc.external_dirty_slot_pages",
-            m.len(),
-            map_bytes(&m) + inner,
-        ));
-    });
-    REMEMBERED_SET.with(|s| {
-        let s = s.borrow();
-        rows.push(("gc.remembered_set", s.len(), set_bytes(&s)));
-    });
-    EVER_DIRTY_OLD_PAGES.with(|s| {
-        let s = s.borrow();
-        rows.push(("gc.ever_dirty_old_pages", s.len(), set_bytes(&s)));
-    });
-    rows
-}
+pub(crate) mod census_rows;
